Assignment

+++ Crux29Dec/src/Lec3/Temp_conv.java
@@ -11,7 +11,7 @@ public class Temp_conv {
 		int max_t = scn.nextInt();
 		int step = scn.nextInt();
 		while(min_t<=max_t) {
-			int c = (int)((float)(5.0/9)*(min_t - 32));
+			int c = (int)((5.0/9)*(min_t - 32));
 			System.out.println(min_t + "\t" + c);
 			min_t+=step;
 		}
